// src/types/data.ts
export type MovieCardData = {
  id: number;
  thumbnail: string;
  title: string;
  year: number;
  rating: number;
};

export type MovieDetailData = {
  backdropImage: string;
  posterImage: string;
  title: string;
  rating: number;
  runtime: number;
  tagline: string;
  releaseYear: string;
  overview: string;
};